* minggo/metal-support:
  clipping node support
  fix compiling error
  fix batch issue
  add draw test back

# Conflicts:
#	cocos/renderer/ccShaders.cpp
#	cocos/renderer/ccShaders.h

+++ cocos/2d/CCAutoPolygon.cpp
@@ -644,7 +644,7 @@ TrianglesCommand::Triangles AutoPolygon::triangulate(const std::vector<Vec2>& po
 
     // Triangles should really use std::vector and not arrays for verts and indices. 
     // Then the above memcpy would not be necessary
-    TrianglesCommand::Triangles triangles = { vertsBuf, indicesBuf, static_cast<int>(verts.size()), static_cast<int>(indices.size()) };
+    TrianglesCommand::Triangles triangles = { vertsBuf, indicesBuf, verts.size(), indices.size() };
     return triangles;
 }
